app: return 0.0..1.0 from gimp_histogram_get_mean(), median(), std_dev()
Display the float values in the histogram dockable and add a compat
hack to the gimp_histogram() PDB wrapper.

+++ app/core/gimphistogram.h
@@ -72,7 +72,7 @@ gdouble         gimp_histogram_get_mean      (GimpHistogram        *histogram,
                                               GimpHistogramChannel  channel,
                                               gint                  start,
                                               gint                  end);
-gint            gimp_histogram_get_median    (GimpHistogram        *histogram,
+gdouble         gimp_histogram_get_median    (GimpHistogram        *histogram,
                                               GimpHistogramChannel  channel,
                                               gint                  start,
                                               gint                  end);
